+++ app/worker.php
@@ -37,9 +37,7 @@ use Utopia\Queue\Message;
 use Utopia\Queue\Publisher;
 use Utopia\Queue\Server;
 use Utopia\Registry\Registry;
-use Utopia\Storage\Device;
 use Utopia\System\System;
-use Utopia\Telemetry\Adapter as Telemetry;
 use Utopia\Telemetry\Adapter\None as NoTelemetry;
 
 Authorization::disable();
@@ -339,26 +337,38 @@ Server::setResource('pools', function (Registry $register) {
 
 Server::setResource('telemetry', fn () => new NoTelemetry());
 
-Server::setResource('deviceForFunctions', function (Document $project, Telemetry $telemetry) {
-    return new Device\Telemetry($telemetry, getDevice(APP_STORAGE_FUNCTIONS . '/app-' . $project->getId()));
-}, ['project', 'telemetry']);
+Server::setResource('deviceForSites', function (Document $project) {
+    return getDevice(APP_STORAGE_SITES . '/app-' . $project->getId());
+}, ['project']);
 
-Server::setResource('deviceForFiles', function (Document $project, Telemetry $telemetry) {
-    return new Device\Telemetry($telemetry, getDevice(APP_STORAGE_UPLOADS . '/app-' . $project->getId()));
-}, ['project', 'telemetry']);
+Server::setResource('deviceForImports', function (Document $project) {
+    return getDevice(APP_STORAGE_IMPORTS . '/app-' . $project->getId());
+}, ['project']);
 
-Server::setResource('deviceForBuilds', function (Document $project, Telemetry $telemetry) {
-    return new Device\Telemetry($telemetry, getDevice(APP_STORAGE_BUILDS . '/app-' . $project->getId()));
-}, ['project', 'telemetry']);
+Server::setResource('deviceForFunctions', function (Document $project) {
+    return getDevice(APP_STORAGE_FUNCTIONS . '/app-' . $project->getId());
+}, ['project']);
 
-Server::setResource('deviceForCache', function (Document $project, Telemetry $telemetry) {
-    return new Device\Telemetry($telemetry, getDevice(APP_STORAGE_CACHE . '/app-' . $project->getId()));
-}, ['project', 'telemetry']);
+Server::setResource('deviceForFiles', function (Document $project) {
+    return getDevice(APP_STORAGE_UPLOADS . '/app-' . $project->getId());
+}, ['project']);
+
+Server::setResource('deviceForBuilds', function (Document $project) {
+    return getDevice(APP_STORAGE_BUILDS . '/app-' . $project->getId());
+}, ['project']);
+
+Server::setResource('deviceForCache', function (Document $project) {
+    return getDevice(APP_STORAGE_CACHE . '/app-' . $project->getId());
+}, ['project']);
 
 Server::setResource(
     'isResourceBlocked',
     fn () => fn (Document $project, string $resourceType, ?string $resourceId) => false
 );
+
+Server::setResource('plan', function (array $plan = []) {
+    return [];
+});
 
 Server::setResource('certificates', function () {
     $email = System::getEnv('_APP_EMAIL_CERTIFICATES', System::getEnv('_APP_SYSTEM_SECURITY_EMAIL_ADDRESS'));
@@ -370,7 +380,7 @@ Server::setResource('certificates', function () {
 });
 
 Server::setResource('logError', function (Registry $register, Document $project) {
-    return function (Throwable $error, string $namespace, string $action, ?array $extras) use ($register, $project) {
+    return function (Throwable $error, string $namespace, string $action, ?array $extras = null) use ($register, $project) {
         $logger = $register->get('logger');
 
         if ($logger) {
@@ -392,7 +402,7 @@ Server::setResource('logError', function (Registry $register, Document $project)
             $log->addExtra('trace', $error->getTraceAsString());
 
 
-            foreach ($extras as $key => $value) {
+            foreach (($extras ?? []) as $key => $value) {
                 $log->addExtra($key, $value);
             }