+++ src/Appwrite/Platform/Workers/Certificates.php
@@ -8,7 +8,7 @@ use Appwrite\Event\Func;
 use Appwrite\Event\Mail;
 use Appwrite\Event\Realtime;
 use Appwrite\Event\Webhook;
-use Appwrite\Network\Validator\CNAME;
+use Appwrite\Network\Validator\DNS;
 use Appwrite\Template\Template;
 use Appwrite\Utopia\Response\Model\Rule;
 use Exception;
@@ -28,6 +28,8 @@ use Utopia\Logger\Log;
 use Utopia\Platform\Action;
 use Utopia\Queue\Message;
 use Utopia\System\System;
+use Utopia\Validator\AnyOf;
+use Utopia\Validator\IP;
 
 class Certificates extends Action
 {
@@ -52,10 +54,8 @@ class Certificates extends Action
             ->inject('queueForRealtime')
             ->inject('log')
             ->inject('certificates')
-            ->callback(
-                fn (Message $message, Database $dbForPlatform, Mail $queueForMails, Event $queueForEvents, Webhook $queueForWebhooks, Func $queueForFunctions, Realtime $queueForRealtime, Log $log, CertificatesAdapter $certificates) =>
-                    $this->action($message, $dbForPlatform, $queueForMails, $queueForEvents, $queueForWebhooks, $queueForFunctions, $queueForRealtime, $log, $certificates)
-            );
+            ->inject('plan')
+            ->callback([$this, 'action']);
     }
 
     /**
@@ -72,8 +72,18 @@ class Certificates extends Action
      * @throws Throwable
      * @throws \Utopia\Database\Exception
      */
-    public function action(Message $message, Database $dbForPlatform, Mail $queueForMails, Event $queueForEvents, Webhook $queueForWebhooks, Func $queueForFunctions, Realtime $queueForRealtime, Log $log, CertificatesAdapter $certificates): void
-    {
+    public function action(
+        Message $message,
+        Database $dbForPlatform,
+        Mail $queueForMails,
+        Event $queueForEvents,
+        Webhook $queueForWebhooks,
+        Func $queueForFunctions,
+        Realtime $queueForRealtime,
+        Log $log,
+        CertificatesAdapter $certificates,
+        array $plan
+    ): void {
         $payload = $message->getPayload() ?? [];
 
         if (empty($payload)) {
@@ -86,7 +96,7 @@ class Certificates extends Action
 
         $log->addTag('domain', $domain->get());
 
-        $this->execute($domain, $dbForPlatform, $queueForMails, $queueForEvents, $queueForWebhooks, $queueForFunctions, $queueForRealtime, $log, $certificates, $skipRenewCheck);
+        $this->execute($domain, $dbForPlatform, $queueForMails, $queueForEvents, $queueForWebhooks, $queueForFunctions, $queueForRealtime, $log, $certificates, $skipRenewCheck, $plan);
     }
 
     /**
@@ -98,12 +108,24 @@ class Certificates extends Action
      * @param Realtime $queueForRealtime
      * @param CertificatesAdapter $certificates
      * @param bool $skipRenewCheck
+     * @param array $plan
      * @return void
      * @throws Throwable
      * @throws \Utopia\Database\Exception
      */
-    private function execute(Domain $domain, Database $dbForPlatform, Mail $queueForMails, Event $queueForEvents, Webhook $queueForWebhooks, Func $queueForFunctions, Realtime $queueForRealtime, Log $log, CertificatesAdapter $certificates, bool $skipRenewCheck = false): void
-    {
+    private function execute(
+        Domain $domain,
+        Database $dbForPlatform,
+        Mail $queueForMails,
+        Event $queueForEvents,
+        Webhook $queueForWebhooks,
+        Func $queueForFunctions,
+        Realtime $queueForRealtime,
+        Log $log,
+        CertificatesAdapter $certificates,
+        bool $skipRenewCheck = false,
+        array $plan = []
+    ): void {
         /**
          * 1. Read arguments and validate domain
          * 2. Get main domain
@@ -184,7 +206,7 @@ class Certificates extends Action
             $certificate->setAttribute('renewDate', DateTime::now());
 
             // Send email to security email
-            $this->notifyError($domain->get(), $e->getMessage(), $attempts, $queueForMails);
+            $this->notifyError($domain->get(), $e->getMessage(), $attempts, $queueForMails, $plan);
 
             throw $e;
         } finally {
@@ -212,8 +234,16 @@ class Certificates extends Action
      * @throws Conflict
      * @throws Structure
      */
-    private function saveCertificateDocument(string $domain, Document $certificate, bool $success, Database $dbForPlatform, Event $queueForEvents, Webhook $queueForWebhooks, Func $queueForFunctions, Realtime $queueForRealtime): void
-    {
+    private function saveCertificateDocument(
+        string $domain,
+        Document $certificate,
+        bool $success,
+        Database $dbForPlatform,
+        Event $queueForEvents,
+        Webhook $queueForWebhooks,
+        Func $queueForFunctions,
+        Realtime $queueForRealtime
+    ): void {
         // Check if update or insert required
         $certificateDocument = $dbForPlatform->findOne('certificates', [Query::equal('domain', [$domain])]);
         if (!$certificateDocument->isEmpty()) {
@@ -266,22 +296,39 @@ class Certificates extends Action
         }
 
         if (!$isMainDomain) {
-            // TODO: Would be awesome to also support A/AAAA records here. Maybe dry run?
-            // Validate if domain target is properly configured
-            $target = new Domain(System::getEnv('_APP_DOMAIN_TARGET', ''));
+            $validationStart = \microtime(true);
 
-            if (!$target->isKnown() || $target->isTest()) {
-                throw new Exception('Unreachable CNAME target (' . $target->get() . '), please use a domain with a public suffix.');
+            $validators = [];
+            $targetCNAME = new Domain(System::getEnv('_APP_DOMAIN_TARGET_CNAME', ''));
+            if ($targetCNAME->isKnown() && !$targetCNAME->isTest()) {
+                $validators[] = new DNS($targetCNAME->get(), DNS::RECORD_CNAME);
+            }
+            if ((new IP(IP::V4))->isValid(System::getEnv('_APP_DOMAIN_TARGET_A', ''))) {
+                $validators[] = new DNS(System::getEnv('_APP_DOMAIN_TARGET_A', ''), DNS::RECORD_A);
+            }
+            if ((new IP(IP::V6))->isValid(System::getEnv('_APP_DOMAIN_TARGET_AAAA', ''))) {
+                $validators[] = new DNS(System::getEnv('_APP_DOMAIN_TARGET_AAAA', ''), DNS::RECORD_AAAA);
+            }
+
+            // Validate if domain target is properly configured
+            if (empty($validators)) {
+                throw new Exception('At least one of domain targets environment variable must be configured.');
             }
 
             // Verify domain with DNS records
-            $validationStart = \microtime(true);
-            $validator = new CNAME($target->get());
+            $validator = new AnyOf($validators, AnyOf::TYPE_STRING);
             if (!$validator->isValid($domain->get())) {
                 $log->addExtra('dnsTiming', \strval(\microtime(true) - $validationStart));
                 $log->addTag('dnsDomain', $domain->get());
 
-                $error = $validator->getLogs();
+                $errors = [];
+                foreach ($validators as $validator) {
+                    if (!empty($validator->getLogs())) {
+                        $errors[] = $validator->getLogs();
+                    }
+                }
+
+                $error = \implode("\n", $errors);
                 $log->addExtra('dnsResponse', \is_array($error) ? \json_encode($error) : \strval($error));
 
                 throw new Exception('Failed to verify domain DNS records.');
@@ -299,10 +346,11 @@ class Certificates extends Action
      * @param string $errorMessage Verbose error message
      * @param int $attempt How many times it failed already
      * @param Mail $queueForMails
+     * @param array $plan
      * @return void
      * @throws Exception
      */
-    private function notifyError(string $domain, string $errorMessage, int $attempt, Mail $queueForMails): void
+    private function notifyError(string $domain, string $errorMessage, int $attempt, Mail $queueForMails, array $plan): void
     {
         // Log error into console
         Console::warning('Cannot renew domain (' . $domain . ') on attempt no. ' . $attempt . ' certificate: ' . $errorMessage);
@@ -314,6 +362,15 @@ class Certificates extends Action
         $template->setParam('{{domain}}', $domain);
         $template->setParam('{{error}}', \nl2br($errorMessage));
         $template->setParam('{{attempts}}', $attempt);
+
+        $template->setParam('{{logoUrl}}', $plan['logoUrl'] ?? APP_EMAIL_LOGO_URL);
+        $template->setParam('{{accentColor}}', $plan['accentColor'] ?? APP_EMAIL_ACCENT_COLOR);
+        $template->setParam('{{twitterUrl}}', $plan['twitterUrl'] ?? APP_SOCIAL_TWITTER);
+        $template->setParam('{{discordUrl}}', $plan['discordUrl'] ?? APP_SOCIAL_DISCORD);
+        $template->setParam('{{githubUrl}}', $plan['githubUrl'] ?? APP_SOCIAL_GITHUB_APPWRITE);
+        $template->setParam('{{termsUrl}}', $plan['termsUrl'] ?? APP_EMAIL_TERMS_URL);
+        $template->setParam('{{privacyUrl}}', $plan['privacyUrl'] ?? APP_EMAIL_PRIVACY_URL);
+
         $body = $template->render();
 
         $emailVariables = [
@@ -345,8 +402,16 @@ class Certificates extends Action
      *
      * @return void
      */
-    private function updateDomainDocuments(string $certificateId, string $domain, bool $success, Database $dbForPlatform, Event $queueForEvents, Webhook $queueForWebhooks, Func $queueForFunctions, Realtime $queueForRealtime): void
-    {
+    private function updateDomainDocuments(
+        string $certificateId,
+        string $domain,
+        bool $success,
+        Database $dbForPlatform,
+        Event $queueForEvents,
+        Webhook $queueForWebhooks,
+        Func $queueForFunctions,
+        Realtime $queueForRealtime
+    ): void {
         // TODO: @christyjacob remove once we migrate the rules in 1.7.x
         if (System::getEnv('_APP_RULES_FORMAT') === 'md5') {
             $rule = $dbForPlatform->getDocument('rules', md5($domain));